Add multiple QOL changes
This commit adds convenience functions to encode `Readble` and `Writeable`
types directly with big or little endian encoding.

This also adds the option to specify the supported endianness on `Readable`
types. This can be checked with `Self::supports::<E>()`.

The next commit will also introduce this functionality for `Writeable`
types.