feat: Add different modes of reading a string
This commit adds the initial work needed to support a more generic
way of reading strings from the reader, sized and terminated with
an optional maximum length. This maximum length can be used to
enforce protocol-specific limits on string lengths.

It also moves the code into the crates subfolder to contain all
Rust code in a single folder.